187: cull based on repeated gaps
gaps that are repeated multiple times are special.

if they are repeated more times than their length, there is no way to
choose one element from each set.

if they are repeated the same number of times as their elements, we know
that we must choose their elements from them, so we can remove them and
the elements they contain from other gaps.

if when removing elements, we notice that we have a shared number in the
special set, then we know we cant select one from each successfully. And
if we end up removing more gaps completely than we expected to, then we
know that another gap was fully covered by the set of elements in
special sets (and thus would not be able to uniquely choose one of its
own elements after we were forced to pick from the special sets.